Refactoring Component name
