+++ frontend/src/components/debate/TopicListNavBar.js
@@ -4,7 +4,7 @@ import { Nav, Navbar, NavDropdown } from "react-bootstrap";
 import { routerEndPoint } from "../common/Constants";
 import { useAuth, logout } from "../authenticated/auth";
 import { getCurrentUser } from "../authenticated/AuthService";
-import LogoutButton from "../sign/SignOut";
+import LogoutButton from "../sign/LogoutButton";
 import ManagerComponent from "../management/ManagerComponent";
 import "bootstrap/dist/css/bootstrap.css";
